refactoring in TS way- "namespace"
