+++ src/decorator/autobind.ts
@@ -1,0 +1,19 @@
+namespace App {
+  //it's autobind decorator. refer to the last section's code for details
+  export function AutoBindThis(
+    _: any,
+    _2: any,
+    descriptor: PropertyDescriptor
+  ) {
+    const originalMethod = descriptor.value;
+    const adjustedDescriptor: PropertyDescriptor = {
+      configurable: true,
+      enumerable: false,
+      get() {
+        const boundFn = originalMethod.bind(this);
+        return boundFn;
+      },
+    };
+    return adjustedDescriptor;
+  }
+}
