+++ src/models/drag-drop-interfaces.ts
@@ -1,0 +1,15 @@
+// Drag and drop interfaces
+
+//namespace is a TS feature
+namespace App {
+  // 'export' make the Draggable available outside this file
+  export interface Draggable {
+    dragStartHandler(event: DragEvent): void;
+    dragEndHandler(event: DragEvent): void;
+  }
+  export interface DragTarget {
+    dragOverHandler(event: DragEvent): void;
+    dropHandler(event: DragEvent): void;
+    dragLeaveHandler(event: DragEvent): void;
+  }
+}
